Extract and export the `publicKey` serializer
