+++ packages/umi-serializer-beet/src/index.ts
@@ -1,3 +1,4 @@
 export * from './createBeetSerializer';
 export * from './errors';
 export * from './plugin';
+export * from './pubkey';
