+++ packages/umi-serializer-beet/test/publicKey.test.ts
@@ -1,11 +1,9 @@
 import test from 'ava';
 import { base16, publicKey as toPublicKey } from '@metaplex-foundation/umi';
-import { createBeetSerializer } from '../src';
 import { s, d } from './_helpers';
+import { publicKey } from '../src/pubkey';
 
 test('serialization', (t) => {
-  const { publicKey } = createBeetSerializer();
-
   const keyA = toPublicKey('4HM9LW2rm3SR2ZdBiFK3D21ENmQWpqEJEhx1nfgcC3r9');
   const keyABytes = base16.deserialize(keyA.bytes)[0];
   s(t, publicKey(), keyA, keyABytes);
@@ -23,8 +21,6 @@ test('serialization', (t) => {
 });
 
 test('deserialization', (t) => {
-  const { publicKey } = createBeetSerializer();
-
   const keyA = toPublicKey('4HM9LW2rm3SR2ZdBiFK3D21ENmQWpqEJEhx1nfgcC3r9');
   const keyABytes = base16.deserialize(keyA.bytes)[0];
   d(t, publicKey(), keyABytes, keyA, 32);
@@ -35,13 +31,11 @@ test('deserialization', (t) => {
 });
 
 test('description', (t) => {
-  const { publicKey } = createBeetSerializer();
   t.is(publicKey().description, 'publicKey');
   t.is(publicKey({ description: 'My publicKey' }).description, 'My publicKey');
 });
 
 test('sizes', (t) => {
-  const { publicKey } = createBeetSerializer();
   t.is(publicKey().fixedSize, 32);
   t.is(publicKey().maxSize, 32);
 });
