+++ packages/umi-serializer-data-view/src/index.ts
@@ -1,3 +1,4 @@
 export * from './createDataViewSerializer';
 export * from './errors';
 export * from './plugin';
+export * from './pubkey';
